Keep image capability tests below size limit
